+++ src/pager.rs
@@ -1,5 +1,7 @@
 //! [`Pager`] allows the user to remotely wait for a desired resource.
 
+#![deny(unsafe_code)]
+
 use std::cell::UnsafeCell;
 use std::marker::{PhantomData, PhantomPinned};
 use std::pin::Pin;
@@ -14,7 +16,7 @@ use crate::wait_queue::{Entry, PinnedEntry, WaitQueue};
 #[derive(Debug, Default)]
 pub struct Pager<'s, S: SyncResult> {
     /// The wait queue for the [`Pager`].
-    wait_queue: UnsafeCell<Option<WaitQueue>>,
+    wait_queue: UnsafeCell<WaitQueue>,
     /// The [`Pager`] cannot outlive the associated synchronization primitive.
     _phantom: PhantomData<&'s S>,
     /// The [`Pager`] cannot be unpinned.
@@ -63,35 +65,8 @@ impl<'s, S: SyncResult> Pager<'s, S> {
     /// assert_eq!(gate.open().1, 1);
     /// ```
     #[inline]
-    pub fn is_registered(&self) -> bool {
-        self.wait_queue().is_some()
-    }
-
-    /// Returns `true` if the [`Pager`] can only be polled synchronously.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::pin::pin;
-    ///
-    /// use saa::{Gate, Pager};
-    /// use saa::gate::State;
-    ///
-    /// let gate = Gate::default();
-    ///
-    /// let mut pinned_pager = pin!(Pager::default());
-    ///
-    /// assert!(gate.register_pager(&mut pinned_pager, true));
-    /// assert!(pinned_pager.is_sync());
-    ///
-    /// assert_eq!(gate.open().1, 1);
-    ///
-    /// assert_eq!(pinned_pager.poll_sync(), Ok(State::Open));
-    /// assert!(!pinned_pager.is_sync());
-    /// ```
-    #[inline]
-    pub fn is_sync(&self) -> bool {
-        self.wait_queue().is_some_and(|w| w.entry().is_sync())
+    pub fn is_registered(self: &mut Pin<&mut Pager<'s, S>>) -> bool {
+        self.wait_queue().is_enqueued()
     }
 
     /// Waits for the desired resource to become available asynchronously.
@@ -118,17 +93,18 @@ impl<'s, S: SyncResult> Pager<'s, S> {
     /// ```
     #[inline]
     pub async fn poll_async(self: &mut Pin<&mut Pager<'s, S>>) -> S::Result {
-        let Some(wait_queue) = self.wait_queue() else {
-            return S::to_result(0, Some(Error::NotRegistered));
-        };
+        if !self.is_registered() {
+            let Some(result) = self.wait_queue().entry().try_acknowledge_result() else {
+                return S::to_result(0, Some(Error::NotRegistered));
+            };
+            return S::to_result(result, None);
+        }
+        let wait_queue = self.wait_queue();
         let pinned_entry = PinnedEntry(Pin::new(wait_queue.entry()));
         let result = pinned_entry.await;
         if result == Entry::ERROR_WRONG_MODE {
             return S::to_result(result, Some(Error::WrongMode));
         }
-        self.with_wait_queue_mut(|e| {
-            *e = None;
-        });
         S::to_result(result, None)
     }
 
@@ -154,18 +130,17 @@ impl<'s, S: SyncResult> Pager<'s, S> {
     /// ```
     #[inline]
     pub fn poll_sync(self: &mut Pin<&mut Pager<'s, S>>) -> S::Result {
-        self.with_wait_queue_mut(|w| {
-            let Some(wait_queue) = w else {
+        if !self.is_registered() {
+            let Some(result) = self.wait_queue().entry().try_acknowledge_result() else {
                 return S::to_result(0, Some(Error::NotRegistered));
             };
-            let entry = wait_queue.entry();
-            let result = entry.poll_result_sync();
-            if result == Entry::ERROR_WRONG_MODE {
-                return S::to_result(0, Some(Error::WrongMode));
-            }
-            *w = None;
-            S::to_result(result, None)
-        })
+            return S::to_result(result, None);
+        }
+        let result = self.wait_queue().entry().poll_result_sync();
+        if result == Entry::ERROR_WRONG_MODE {
+            return S::to_result(result, Some(Error::WrongMode));
+        }
+        S::to_result(result, None)
     }
 
     /// Tries to get the result.
@@ -188,51 +163,25 @@ impl<'s, S: SyncResult> Pager<'s, S> {
     /// assert_eq!(gate.open().1, 1);
     ///
     /// assert_eq!(pinned_pager.try_poll(), Ok(State::Open));
-    /// assert_eq!(pinned_pager.poll_sync(), Ok(State::Open));
     /// ```
     #[inline]
-    pub fn try_poll(&self) -> S::Result {
-        self.with_wait_queue(|w| {
-            let Some(wait_queue) = w else {
+    pub fn try_poll(self: &mut Pin<&mut Pager<'s, S>>) -> S::Result {
+        if !self.is_registered() {
+            let Some(result) = self.wait_queue().entry().try_acknowledge_result() else {
                 return S::to_result(0, Some(Error::NotRegistered));
             };
-
-            let entry = wait_queue.entry();
-            if let Some(result) = entry.try_acknowledge_result() {
-                S::to_result(result, None)
-            } else {
-                S::to_result(0, Some(Error::NotReady))
-            }
-        })
+            return S::to_result(result, None);
+        }
+        if let Some(result) = self.wait_queue().entry().try_acknowledge_result() {
+            S::to_result(result, None)
+        } else {
+            S::to_result(0, Some(Error::NotReady))
+        }
     }
 
     /// Returns a reference to the wait queue entry.
     #[inline]
-    pub(crate) fn wait_queue(&self) -> Option<Pin<&WaitQueue>> {
-        unsafe {
-            (*self.wait_queue.get())
-                .as_ref()
-                .map(|w| Pin::new_unchecked(w))
-        }
-    }
-
-    /// Sets the wait queue entry.
-    ///
-    /// The user must make sure that the `self` is exclusively owned.
-    #[inline]
-    pub(crate) fn set_wait_queue(&self) {
-        self.with_wait_queue_mut(|e| e.replace(WaitQueue::default()));
-    }
-
-    /// Returns a reference to the wait queue.
-    #[inline]
-    fn with_wait_queue<R, F: FnOnce(&Option<WaitQueue>) -> R>(&self, f: F) -> R {
-        unsafe { f(&(*self.wait_queue.get())) }
-    }
-
-    /// Returns a reference to the wait queue.
-    #[inline]
-    fn with_wait_queue_mut<R, F: FnOnce(&mut Option<WaitQueue>) -> R>(&self, f: F) -> R {
-        unsafe { f(&mut (*self.wait_queue.get())) }
+    pub(crate) fn wait_queue(&self) -> Pin<&WaitQueue> {
+        WaitQueue::pinned_wait_queue(self.wait_queue.get())
     }
 }