+++ src/sync_primitive.rs
@@ -56,7 +56,7 @@ pub(crate) trait SyncPrimitive: Sized {
             .is_ok()
         {
             // The entry cannot be dropped until the result is acknowledged.
-            wait_queue.entry().enqueued();
+            wait_queue.entry().set_enqueued();
             begin_wait();
             None
         } else {